Move callbackHelpers to separate file.  Start taking export method from container author
