+++ examples/utils/migration-tools/src/migrator/migratorEntryPointPiece.ts
@@ -3,19 +3,17 @@
  * Licensed under the MIT License.
  */
 
-import type { IContainer } from "@fluidframework/container-definitions/internal";
 import type { IContainerRuntime } from "@fluidframework/container-runtime-definitions/internal";
 import type { FluidObject } from "@fluidframework/core-interfaces";
 
 import type { IEntryPointPiece } from "../compositeRuntime/index.js";
 import { MigrationToolFactory, type IMigrationTool } from "../migrationTool/index.js";
-import { type ISimpleLoader } from "../simpleLoader/index.js";
 
-import type { DataTransformationCallback, IMigratableModel } from "./interfaces.js";
 import {
-	Migrator,
+	type ExportDataCallback,
 	type LoadSourceContainerCallback,
 	type MigrationCallback,
+	Migrator,
 } from "./migrator.js";
 
 const migratorEntryPointPieceName = "getMigrator";
@@ -38,101 +36,43 @@ async function getDataStoreEntryPoint(
 }
 
 /**
- * Helper function for casting the container's entrypoint to the expected type.  Does a little extra
- * type checking for added safety.
- */
-const getModelFromContainer = async <ModelType>(container: IContainer): Promise<ModelType> => {
-	const entryPoint = (await container.getEntryPoint()) as {
-		model: ModelType;
-	};
-
-	// If the user tries to use this with an incompatible container runtime, we want to give them
-	// a comprehensible error message.  So distrust the type by default and do some basic type checking.
-	if (typeof entryPoint.model !== "object") {
-		throw new TypeError("Incompatible container runtime: doesn't provide model");
-	}
-
-	return entryPoint.model;
-};
-
-/**
- * Make a typical migration callback.
+ * Create the entry point piece.  To be called by the container author, who can provide appropriate access to the
+ * container data.
  * @alpha
  */
-export const makeMigrationCallback = (
-	loader: ISimpleLoader,
-	dataTransformationCallback?: DataTransformationCallback | undefined,
-): MigrationCallback => {
-	const migrationCallback = async (
-		version: string,
-		exportedData: unknown,
-	): Promise<unknown> => {
-		const detachedContainer = await loader.createDetached(version);
-		const destinationModel = await getModelFromContainer<IMigratableModel>(
-			detachedContainer.container,
-		);
-		// TODO: Is there a reasonable way to validate at proposal time whether we'll be able to get the
-		// exported data into a format that the new model can import?  If we can determine it early, then
-		// clients with old MigratableModelLoaders can use that opportunity to dispose early and try to get new
-		// MigratableModelLoaders.
-		// TODO: Error paths in case the format isn't ingestible.
-		let transformedData: unknown;
-		if (destinationModel.supportsDataFormat(exportedData)) {
-			// If the migrated model already supports the data format, go ahead with the migration.
-			transformedData = exportedData;
-		} else if (dataTransformationCallback !== undefined) {
-			// Otherwise, try using the dataTransformationCallback if provided to get the exported data into
-			// a format that we can import.
-			transformedData = await dataTransformationCallback(
-				exportedData,
-				destinationModel.version,
-			);
-		}
-		await destinationModel.importData(transformedData);
-		const newContainerId = await detachedContainer.attach();
-		detachedContainer.container.dispose();
-		return newContainerId;
+export const makeMigratorEntryPointPiece = (
+	exportDataCallback: ExportDataCallback,
+): IEntryPointPiece => {
+	return {
+		name: migratorEntryPointPieceName,
+		registryEntries: [[migrationToolRegistryKey, Promise.resolve(migrationToolFactory)]],
+		onCreate: async (runtime: IContainerRuntime): Promise<void> => {
+			const migrationTool = await runtime.createDataStore(migrationToolRegistryKey);
+			await migrationTool.trySetAlias(migrationToolId);
+		},
+		onLoad: async (runtime: IContainerRuntime): Promise<void> => {
+			// Force the MigrationTool to instantiate in all cases.  The PactMap it uses must be loaded and running in
+			// order to respond with accept ops, and without this call the MigrationTool won't be instantiated on the
+			// summarizer client.
+			await getDataStoreEntryPoint(runtime, migrationToolId);
+		},
+		createPiece: async (runtime: IContainerRuntime): Promise<FluidObject> => {
+			return async (
+				loadSourceContainerCallback: LoadSourceContainerCallback,
+				migrationCallback: MigrationCallback,
+			) => {
+				const migrationTool = (await getDataStoreEntryPoint(
+					runtime,
+					migrationToolId,
+				)) as IMigrationTool;
+				const migrator = new Migrator(
+					migrationTool,
+					loadSourceContainerCallback,
+					exportDataCallback,
+					migrationCallback,
+				);
+				return migrator;
+			};
+		},
 	};
-	return migrationCallback;
-};
-
-/**
- * @alpha
- */
-export const migratorEntryPointPiece: IEntryPointPiece = {
-	name: migratorEntryPointPieceName,
-	registryEntries: [[migrationToolRegistryKey, Promise.resolve(migrationToolFactory)]],
-	onCreate: async (runtime: IContainerRuntime): Promise<void> => {
-		const migrationTool = await runtime.createDataStore(migrationToolRegistryKey);
-		await migrationTool.trySetAlias(migrationToolId);
-	},
-	onLoad: async (runtime: IContainerRuntime): Promise<void> => {
-		// Force the MigrationTool to instantiate in all cases.  The PactMap it uses must be loaded and running in
-		// order to respond with accept ops, and without this call the MigrationTool won't be instantiated on the
-		// summarizer client.
-		await getDataStoreEntryPoint(runtime, migrationToolId);
-	},
-	createPiece: async (runtime: IContainerRuntime): Promise<FluidObject> => {
-		const exportDataCallback = async (container: IContainer): Promise<unknown> => {
-			// TODO: verify IMigratableModel
-			const exportModel = await getModelFromContainer<IMigratableModel>(container);
-			return exportModel.exportData();
-		};
-		return async (
-			loadSourceContainerCallback: LoadSourceContainerCallback,
-			migrationCallback: MigrationCallback,
-		) => {
-			const migrationTool = (await getDataStoreEntryPoint(
-				runtime,
-				migrationToolId,
-			)) as IMigrationTool;
-			const migrator = new Migrator(
-				migrationTool,
-				loadSourceContainerCallback,
-				exportDataCallback,
-				migrationCallback,
-			);
-			return migrator;
-		};
-	},
 };
