+++ examples/version-migration/separate-container/src/modelVersion1/containerCode.ts
@@ -5,10 +5,12 @@
 
 import {
 	CompositeEntryPoint,
+	type IMigratableModel,
 	loadCompositeRuntime,
-	migratorEntryPointPiece,
+	makeMigratorEntryPointPiece,
 } from "@fluid-example/migration-tools/internal";
 import type {
+	IContainer,
 	IContainerContext,
 	IRuntime,
 	IRuntimeFactory,
@@ -16,6 +18,31 @@ import type {
 import type { IContainerRuntimeOptions } from "@fluidframework/container-runtime/internal";
 
 import { modelEntryPointPiece } from "./modelEntryPointPiece.js";
+
+/**
+ * Helper function for casting the container's entrypoint to the expected type.  Does a little extra
+ * type checking for added safety.
+ */
+const getModelFromContainer = async <ModelType>(container: IContainer): Promise<ModelType> => {
+	const entryPoint = (await container.getEntryPoint()) as {
+		model: ModelType;
+	};
+
+	// If the user tries to use this with an incompatible container runtime, we want to give them
+	// a comprehensible error message.  So distrust the type by default and do some basic type checking.
+	if (typeof entryPoint.model !== "object") {
+		throw new TypeError("Incompatible container runtime: doesn't provide model");
+	}
+
+	return entryPoint.model;
+};
+
+// TODO: Can/should we access the model more directly than going through the IContainer?
+const exportDataCallback = async (container: IContainer): Promise<unknown> => {
+	// TODO: verify IMigratableModel
+	const exportModel = await getModelFromContainer<IMigratableModel>(container);
+	return exportModel.exportData();
+};
 
 /**
  * @internal
@@ -46,6 +73,7 @@ export class InventoryListContainerRuntimeFactory implements IRuntimeFactory {
 	): Promise<IRuntime> {
 		const compositeEntryPoint = new CompositeEntryPoint();
 		compositeEntryPoint.addEntryPointPiece(modelEntryPointPiece);
+		const migratorEntryPointPiece = makeMigratorEntryPointPiece(exportDataCallback);
 		compositeEntryPoint.addEntryPointPiece(migratorEntryPointPiece);
 		return loadCompositeRuntime(context, existing, compositeEntryPoint, this.runtimeOptions);
 	}
